// src/shared/api/generated/model/profileResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileResponseDtoCertificationsItem } from './profileResponseDtoCertificationsItem.ts';
import type { ProfileResponseDtoIntroVideo } from './profileResponseDtoIntroVideo.ts';
import type { ProfileResponseDtoRating } from './profileResponseDtoRating.ts';
import type { ProfileResponseDtoStatus } from './profileResponseDtoStatus.ts';
import type { ProfileResponseDtoTechniqueVideosItem } from './profileResponseDtoTechniqueVideosItem.ts';

export interface ProfileResponseDto {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  membershipId: string;
  isMe: boolean;
  fullName: string;
  /** @nullable */
  staffTitle: string | null;
  /** @nullable */
  headline: string | null;
  /** @nullable */
  bio: string | null;
  specialties: string[];
  languages: string[];
  status: ProfileResponseDtoStatus;
  /** @nullable */
  reviewNote: string | null;
  hasPublishConsent: boolean;
  certifications: ProfileResponseDtoCertificationsItem[];
  /** @nullable */
  introVideo: ProfileResponseDtoIntroVideo;
  techniqueVideos: ProfileResponseDtoTechniqueVideosItem[];
  /** @nullable */
  rating: ProfileResponseDtoRating;
}